Initial commit
This bot is designed as an experiment for the iDevGames IRC server,
specifically to try and gently funnel political talk into the #politics
channel. The #politics channel itself is private to only those with a
specific role. So the bot has the ability to grant and revoke that role
on the request of users, making it self-service.

The actual act of scanning for political speech is incredibly simple:
there's a list of "political" words and they get a response from the
bot.