Add deny_channels to WordWatcher
WordWatcher would warn on words that should nudge to channels
even when that was mentioned in the target channel, which rather
defeats the purpose. The deny_channels member is used to know
which channels to ignore for those nudges.

Rename CmdrMessageHandler to MysteriousMessageHandler, a missed
part of the rename.

Remove empty watchdog.rs file, it was a previous name for
word_watcher.rs.

Correct a typo in the README which was irritating me.