Track infractions
This uses PickleDB-rs, which is basically just a JSON file on disk. It
was simple and accessible, you can see how simple it was to setup up. I
reckon SQLite3 would be the next step, so that the infracted messages
could be logged for later analysis (eg. did the detection method work
properly or is the bot likely just pissing people off)?