Migrate to async Serenity 0.9.1, Dockerized build
Migrates to Serenity 0.9.1 to take advantage of async. For as low
traffic as iDevGames experiences it's well and truly overkill, but it's
fun nonetheless.

Removes the VerbalMoralityHandler. It was unused, I was late in getting
the configuration in repo to a state consistent with what was actually
running.

Dockerizes the build process. Did you know that if I build on my Pop!_OS
20.10 machine and put that binary on my Ubuntu 20.04 server it does not
   work? Indeed, the glibc version on my laptop is newer and named
differently from the one on the server. This is a symptom of disparate
build and production environments, so by dockerizing the build we can
build and link against an environment that resembles our deployment
environment. We could go further and deploy a docker image itself, but I
think this has been plenty enough adventure for one evening already.