Add counter feature, update VMS
The counter is planned to be used in some Discord-based tomfoolery. I
woke up this morning to find two people had tossed every bad word they
could think of at my bot to give me quite the word list. I'm not sure I
want this on my Github, come to think of it.